code/backend_bot/requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
chromadb
fpdf
SpeechRecognition
pyttsx3
transformers
torch
pyaudio
unstructured
sentence-transformers
python-dotenv
langchain-huggingface
streamlit
pysqlite3-binary
chromadb